refactor: separate EntryList component
